+++ src/components/Catalog.css
@@ -1,0 +1,11 @@
+.catalog-section {
+  background-color: #5D5D5D;
+  padding: 70px;
+}
+
+.product-catalog {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 60px 66px;
+}
