+++ src/components/ProductCard/ProductCard.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap');
-
 .product-card button {
   width: 100px;
   height: 40px;
