+++ src/components/ShoppingCart.css
@@ -1,0 +1,31 @@
+.shopping-bag {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.shopping-bag p {
+  color: #F9FAFD;
+  font-size: 18px;
+}
+
+.shopping-bag button {
+  width: 200px;
+  height: 40px;
+  background-color: #B92822;
+  color: #F9FAFD;
+  border-radius: 25px;
+  border: none;
+  font-weight: 700;
+  font-size: 20px;
+}
+
+.shopping-bag button:hover {
+  background-color: burlywood;
+  transition: all 0.5s;
+}
+
+.shopping-bag button:active {
+  background-color: coral;
+  transition: none;
+}
